Update pets_controller_spec.rb
Changed one test in the pets controller spec to be consistent with itself and the owners controller spec. It was searching for `owner[name]` on line 81, then searching for `owner_name` on line 101. In the owner spec, it was searching for pets[name] in both. Also, this test could be passed with `<input  type="text" id="owner_name" name="owner[name]">`, but that doesn't seem as clean as having the id and name match (especially to make it consistent with the owner edit form).